requirements.txt
Flask==0.12.2
gunicorn==19.7.1
mysql-connector-python-rf==2.2.2 
